update (issue #25): added comment and dynamic rating feature

+++ javabnb/src/main/java/poo/javabnb/Comment.java
@@ -2,10 +2,10 @@ package poo.javabnb;
 
 public class Comment {
     public final Client client;
-    public final float rating;
+    public final int rating;
     public final String message;
     
-    public Comment(Client c, float r, String m){
+    public Comment(Client c, int r, String m){
         this.client = c;
         this.rating = r;
         this.message = m;
